aider: Updated PHP version in Dockerfile to php:8.2-apache-alpine.
# Aider chat conversation:

USER: maybe the image php:8.2-alpine, but i know your knowledge is not up to date, so just go for that

ASSISTANT: To update the PHP version in the Dockerfile, we need to modify the Dockerfile. Here is the change:

```Dockerfile
Dockerfile
<<<<<<< HEAD
FROM php:7.4-apache-alpine
=======
FROM php:8.2-apache-alpine
>>>>>>> updated
```

This change will update the PHP version from 7.4 to 8.2 in the Docker image. Please note that this might introduce breaking changes if the application is not compatible with PHP 8.2. It's recommended to test the application thoroughly after this change.