aider: Fix colorization logic to handle tasks with a deadline in the past.
# Aider chat conversation:

USER: i tested the app and saw, that tasks with due date in the past get not colorized with prio1 color

ASSISTANT: I see. The issue seems to be that the colorization logic only checks if the deadline is within a certain number of days in the future, but it doesn't handle the case where the deadline is in the past.

To fix this, we need to modify the `fetchTasks` function in the `html/index.php` file. We will add a condition to check if the deadline is in the past and, if so, set the color to `prio1`.

Here is the *edit block* for the change:

```php
html/index.php
<<<<<<< HEAD
            $deadline = new DateTime($row['deadline']);
            $now = new DateTime();
            $interval = $now->diff($deadline)->days;
            if ($interval <= $settings['days1']) {
                $color = $settings['prio1'];
            } elseif ($interval <= $settings['days2']) {
                $color = $settings['prio2'];
            } elseif ($interval <= $settings['days3']) {
                $color = $settings['prio3'];
            } else {
                $color = 'black';
            }
=======
            $deadline = new DateTime($row['deadline']);
            $now = new DateTime();
            if ($deadline < $now) {
                $color = $settings['prio1'];
            } else {
                $interval = $now->diff($deadline)->days;
                if ($interval <= $settings['days1']) {
                    $color = $settings['prio1'];
                } elseif ($interval <= $settings['days2']) {
                    $color = $settings['prio2'];
                } elseif ($interval <= $settings['days3']) {
                    $color = $settings['prio3'];
                } else {
                    $color = 'black';
                }
            }
>>>>>>> updated
```

This change will ensure that tasks with a deadline in the past are colorized with the `prio1` color.